pyproject.toml
[project]
name = "mount-linker"
version = "0.1.0"
description = "Automatically link devices on mount to a more convenient location."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "systemd-python>=235",
    "watchdog>=6.0.0",
]
